sonar first commit

+++ recette/src/main/java/com/example/recette/service/impl/EtapeServiceImp.java
@@ -3,6 +3,7 @@ package com.example.recette.service.impl;
 import com.example.recette.bean.Etape;
 import com.example.recette.dao.EtapeDao;
 import com.example.recette.service.facade.EtapeService;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,8 +14,12 @@ public class EtapeServiceImp implements EtapeService {
 
 
     @Override
+    @Transactional
     public int deleteByRef(String ref) {
-        return etapeDao.deleteByRef(ref);
+        if (etapeDao.findByRef(ref) != null) {
+            etapeDao.deleteByRef(ref);
+            return 1;
+        } else return -1;
     }
 
 
@@ -24,9 +29,19 @@ public class EtapeServiceImp implements EtapeService {
         return 1;
     }
 
-
     @Override
+    @Transactional
     public int update(Etape etape) {
-        return 0;
+        Etape existingEtape = etapeDao.findByRef(etape.getRef());
+        if (existingEtape != null) {
+            existingEtape.setDescription(etape.getDescription());
+            existingEtape.setOrdre(etape.getOrdre());
+            existingEtape.setDuree(etape.getDuree());
+            etapeDao.save(existingEtape);
+
+            return 1;
+        } else {
+            return 0;
+        }
     }
 }
