+++ recette/src/main/java/com/example/recette/service/impl/IngredientServiceImpl.java
@@ -1,26 +1,28 @@
 package com.example.recette.service.impl;
 
 import com.example.recette.bean.Ingredient;
-import com.example.recette.bean.Recette;
 import com.example.recette.dao.IngredientDao;
 import com.example.recette.service.facade.IngredientService;
-import com.example.recette.service.facade.RecetteService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class IngredientServiceImpl implements IngredientService {
+
     @Autowired
     private IngredientDao ingredientDao;
 
-
     @Override
+    @Transactional
     public int deleteByRef(String ref) {
-        return ingredientDao.deleteByRef(ref);
+        if (ingredientDao.findByRef(ref) != null) {
+            ingredientDao.deleteByRef(ref);
+            return 1;
+        } else {
+            return -1;
+        }
     }
-
 
     @Override
     public int save(Ingredient ingredient) {
@@ -28,9 +30,19 @@ public class IngredientServiceImpl implements IngredientService {
         return 1;
     }
 
-
     @Override
+    @Transactional
     public int update(Ingredient ingredient) {
-        return 0;
+        Ingredient existingIngredient = ingredientDao.findByRef(ingredient.getRef());
+        if (existingIngredient != null) {
+            existingIngredient.setNom(ingredient.getNom());
+            existingIngredient.setQuantite(ingredient.getQuantite());
+            existingIngredient.setUnite(ingredient.getUnite());
+            ingredientDao.save(existingIngredient);
+
+            return 1;
+        } else {
+            return 0;
+        }
     }
 }
